رفعِ ۶ خطای کامپایلِ بیلدِ ۴۶۵ [skip ci]
بازطراحیِ قدمِ ۱ دو تا خطای نوع/ارجاع جا گذاشته بود که بررسیِ ایستای ایمپورت‌محورِ من
نمی‌گرفتشون:

۱. Theme.kt (۲ خطا): `palette.accent` هنوز صدا زده می‌شد، ولی فیلدِ accent از پالت حذف
   و با goldInk جایگزین شده بود. → palette.goldInk

۲. HomeScreen.kt (۴ خطا): پارامترهای HomeBalanceHero رو Long تعریف کرده بودم ولی کلِ
   پروژه مبلغ‌ها رو Double نگه می‌داره (AccountTransactionEntity.amount، balanceOf()،
   fmt()، countUpAmount() همه Double). → Double

هر دو الگو حالا اسکریپتِ بررسیِ اختصاصیِ خودشون رو دارن (دسترسیِ فیلدِ پالت + نوعِ
پارامترهای پول‌مانند) تا دورِ بعد قبل از مصرفِ سهمیه‌ی CI گرفته بشن.

+++ native-android/app/src/main/kotlin/ir/sadteam/loancalc/ui/home/HomeScreen.kt
@@ -256,9 +256,9 @@ fun HomeScreen(
  */
 @Composable
 private fun HomeBalanceHero(
-    totalBalance: Long,
+    totalBalance: Double,
     accountCount: Int,
-    todaySpend: Long,
+    todaySpend: Double,
     privacyMode: Boolean,
     onClick: () -> Unit,
 ) {
